Handle IMDS errors more gracefully
It's possible that IMDS has gone away for a little while, most likely
due to the software on the other side of the Instance MetaData Service
being updated while the instance itself continues to run. In this case,
we want to be able to know that we have to come back later rather than
continue updating routes and configurations with blank data.

So, this patch sets $? appropriately in get_meta, and then adds error
handling code paths to where it's most needed.

Without this error handling, if you're unlucky enough to hit DHCP renew
while IMDS is briefly offline, ec2-net-utils would treat an empty string
as the list of IPs and thus routes to handle rather than the actual
list. The consequences of this are rather predictable: packets don't go
to the right place.

Additionally, we add a test script that can be run on an instance, that
tests attach/detach works correctly (up to the instance limit of number
of ENIs).

Signed-off-by: Stewart Smith <trawets@amazon.com>