Retry fetching IMDSv2 token
Since IMDS can be unavailable for a short period of time (notably when
the system behind IMDS is undergoing a software update), we must also
handle the case of being unable to fetch the IMDSv2 token for a short
period of time.

Without this retry, ENIs would fail to be attached during this time
period, amongst other problems.

For testing, we can just add an iptables rule temporarily to REJECT
connections to IMDS, then wait a bit before dropping the rule. If the
interface still comes up, our retry works!

Signed-off-by: Stewart Smith <trawets@amazon.com>