Support IP prefix delegation
IP Prefix delegation allows an entire IPv4 and/or IPv6 prefix to be assigned to
an ENI.  Standard VPC anti-spoofing rules apply to delegated prefixes, as they
do to primary or secondary IP addresses associated with an ENI.  This means
that traffic sourced from a given IP address must egress via the ENI with which
that address is associated.  So we extend the existing policy routing rule
construction to insert policy rules for delegated prefixes in addition to
secondary IP addresses.

Assignment of individual addresses within the prefix is out of scope
for this package, and would instead be handled by a local IPAM system.  A
typical such system might provision IP addresses for use by locally running
containers, etc.