Update Provides and Obsoletes entries in spec file
The amazon-ec2-net-utils package is intended to replace the ec2-net-utils
package, so it needs to be installable as a drop-in replacement.  In order to
do that, it needs both a Provides and Obsoletes relationship with that package
in order to ensure that rpm recognizes it as a replacement.

Additionally, we only want to declare a Provides relationship on versions of
ec2-net-utils that whose functionality we actually implement.  An unversioned
Provides entry means we provide all possible versions, including ancient ones
like 0.4-1, which other packages may legitimate want to conflict with.