bugfix: remove all secondary IPs if IMDS says there are none
rewrite_alises contains logic to remove individual secondary IP
addresses from an interface in the event that the list given by IMDS has
changed.  It has a short-circuit conditional that purges all addresses
if IMDS says there aren't any, but this was never firing due to a bug.
The bug was a little bit more subtle than it seemed at first glance.

The original code looked like:

rewrite_aliases() {
  aliases=$(get_secondary_ipv4s)
  if [ ${#aliases[*]} -eq 0 ]; then
    remove_aliases
    return
  fi
...
}

get_secondary_ips provides a potentially empty list of IPs by writing
to stdout.  This IP list is assigned to the aliases variable, which is
then used in an array context in the line

if [ ${#aliases[*]} -eq 0 ]; then

However, because the assignment happened in a scalar context, the
assignment was implicitly made to aliases[0].  Per the bash
documentation, "An array variable is considered set if a subscript has
been assigned a value.  The null string is a valid value."  This means
that even in the case when get_secondary_ips returned an empty string,
the aliases[0] value was still considered set.  So evaluating the length
of the array would always return a nonzero value, making it impossible
to reach the conditional block.

By changing the assignment to use an array context, we avoid assigning
to a specific array element in case get_secondary_ips returns an empty
string.  In that event, the assignment looks like aliases=(), which does
result in an array for which ${#aliases[*]} evaluates to 0.

Tests, which consider "the currently deployed behavior" to be correct,
have been updated to reflect the change in behavior.  In particular, the
tests in which there were no secondary IP addresses needed to be updated
to reflect the fact that the short-circuit condition is being triggered.
In cases where there were secondary IP addresses on an interface, the
behavior was unchanged.