Skip retries when IMDS queries for *-prefix keys fail
The Amazon EC2 Instance Metadata Service returns an HTTP 404 error in
response to queries for ipv4-prefix and ipv6-prefix keys if there are
no delegated prefixes assigned to an ENI.  This makes it impossible
for us to distinguish between an actual error and a legitimate
configuration in which there are simply no prefixes assigned.  Rather
than retry until we get a non-error response that may never actually
arrive, we avoid retries altogether when querying prefix keys.