Skip udev rule for VLAN interface
VLAN interfaces require associations with some parent ethernet device. This can cause an issue with udev, which confuses the physical address of the VLAN interface with the parent interface and creates the wrong device rule.

As such, udev rule generation for VLAN interfaces should be skipped. The user of VLAN interface should create the udev rule themselves if needed.