Add repository documentation
LICENSE reflects the Apache 2.0 license.

NOTICE is provided as attribution text as described in the Apache 2.0
license.

CONTRIBUTING.md describes guidelines for contributing to this project.
Content is based on amazon-ec2-utils document, which is itself based
on the standard Amazon boilerplate but has been modified to describe
the branching strategy used in this project.

CODE_OF_CONDUCT.md describes expectations regarding social behavior
when interacting in the context of this project.