Rename ETCDIR variable to avoid conflicts with dhclient-script
Because ec2net-functions gets sourced into the dhclient-script
process, it runs in an environment shared environment.  The
dhclient-script environment already defines ETCDIR to be the top level
of the dhclient configuration, typically /etc/dhcp.  Our usage of
ETCDIR used it to refer, typically, to /etc.  Use a different variable
name to avoid clashing.