Update tests to reflect changes in DHCP and IMDS call activity
dhclient tests are improved to cover both v4 and v6 leases, as well as
interface configuration and de-configuration.

A number of IMDS calls that previously occurred during ec2ifup are now moved to
protocol-specific activity in the dhclient hooks, and the expected output of
the tests is updated to reflect that.