avoid concurrent networkd reloads
When multiple interfaces are configured simultaneously, such as when
an instance with multiple ENI attachments (re)boots, we want to avoid
issuing `networkctl reload` calls for each one in favor of a single
reload once the configuration has stabilized.  To do that, we use
filesystem tricks to ensure that, when there are multiple concurrent
instances of the setup-policy-routes program running, the last one to
exit is the one that reloads networkd.