bugfix: persistent-net-generator skipped all generation
Commit 1cddb38 ("Ignore interfaces that use unrecognized drivers")
made use of an environment variable that is not available at the time
of processing the kernel udev events.  Consequently, generation of
70-persistent-net.rules was skipped in all conditions, when the intent
was for it to be generated when processing interfaces backed by actual
ENI devices.

We fix this by looking at the DEVPATH key instead of the
ENV{ID_NET_DRIVER} variable, as it has been set at the time of rule
processing, and skipping virtual interfaces (e.g. vlan
sub-interfaces).