always enable the DHCPv4 client
This is a tradeoff.  Currently in an IPv6-only subnet, the DHCPv4
server still distributes a default IPv4 route.  However, traffic to
most destinations ends up being silently dropped, with the only
exceptions being local services such as IMDS and DNS.  Since the IPv6
IMDS endpoint is disabled by default and needs to be explicitly
enabled with an authenticated EC2 API call, access to the IPv4
endpoint is desirable.  Letting the DHCP client set up the routes
unconditionally turns out to be simpler and more reliable than trying
to figure out whether or not we should listen to DHCP or set up our
own routes by hand.

The downside of keeping DHCPv4 enabled is that, in an IPv6-only
subnet, any IPv4 clients attempting to contact external services will
see their traffic blackholed, which will result in timeouts rather
than a fast-fail ENETUNREACH.