Change how we detect support for IPv4 networks
Instead of looking for non 169.254/16 addresses, look explicitly for
the presence of 169.254/16 addresses.  Specifically the differences in
behavior applies to the case when there are no IPv4 addresses on an
interface.  Previously we'd interpret that as a lack of v4 support,
but that's the initial state of an interface before the DHCP
negotiation has completed, so we should bias towards indicating the
presence of v4 support.