bug: ensure only secondary addresses are used when assigning aliases
The `local-ipv4s` IMDS field lists the primary RFC 1918 address on the
first line, with secondary IPv4 addresses on subsequent lines.  When
assigning alises, we want to only consider secondary IPv4 addresses,
so we exclude the primary using `tail -n +2`.  In order to ensure
consistent ordering across invocations, even if IMDS returns secondary
addresses in unspecified order, the list of secondaries is processed
with `sort`.  Prior to this change, we were performing the `sort`
operation before the `tail -n +2` operation, meaning that a secondary
address that sorted before the primary would be mistakenly interpreted
as a primary address, and a secondary address could be mistaken for
the primary address.

See https://github.com/amazonlinux/amazon-ec2-net-utils/issues/63