Skip routing rules for default mac interface
Changed lib.sh setup_interface() to fetch the default mac and the device-
number for the interface being setup from IMDS. Also changed to skip
writing the routing rules table entry for the current interface if it's
device-number is 0 (zero) and it's mac matches the default mac for the
instance. Also fixed a bug in how $ifid is derived from the interface
name by having it also remove uppercase characters from the name.

This change was made to resolve a problem where docker containers could
not communicate with each other or with the host instance. Network
traffic was exiting the container correctly but responses would be
routed elsewhere instead of being sent back to the container.