Improve lockfile handling and perform out-of-band cleanup
Name lock file based on the interface name rather than the PID, so
they're more easily predictable.

Use -o noclobber to ensure we abort if the lock file already exists
when we try to create it.

Implement a cleanup function and have systemd invoke it in the unit
files after execution.  This should normally never find anything, but
in the event that a lock file was somehow leaked, this will ensure
that we can recover.