bug: fix management of IPv4 route configuration file
This addresses a regression in the handling of
/etc/sysconfig/network-scripts/route-* introduced in version 1.6.1
with commit 4b3d2f587 ("Move route file creation to a dhclient enter
hook")

Prior to that commit, ec2ifup would trigger the removal of the route
config file, but otherwise that file would be preserved.  The
regression introduced by that commit was that the route file would be
overwritten any time `ifup` brought the interface up, not only by
`ec2ifup`.

Fixes https://github.com/amazonlinux/amazon-ec2-net-utils/issues/82