Fix a process synchronization issue during ENI removal
When an ENI is present and the refresh timer is running, it is
possible for an ENI removal event to occur during the
refresh-policy-routes job.  When this happens, udev invokes the
'cleanup' task, which should block until it's able to register a
networkd reloader with the `register_networkd_reloader()` function.
Because of an issue with how failed redirects are handled under -o
errexit, the cleanup job was exiting instead of retrying the lock
acquisition.

This change fixes this by disabling -o errexit during the relevant
block in `register_networkd_reloader()`.

It also introduces a 100ms delay between loop iterations and a
terminal condition to catch situations where the lock cannot be
obtained after a long period and is unlikely to ever become available.