Don't install rules for the primary interface
There's an expectation that containers on the docker0 bridge can
communicate with the host on its primary interface's address.  The
practical way to accomplish that, without knowing additional
information about the additional interfaces and/or prefixes involved,
is to avoid creating a rule matching traffic from the primary
interface's addresses and instead let the main table handle that
traffic.

Routing traffic from the primary interface via the main table was in
place prior to version 2.4.0, which regressed.

The tradeoff in fixing this is that traffic that specifically binds to
one of device-number-0's addresses will be routed via the "wrong"
table when trying to reach addresses reachable via a secondary ENI.
The expected behavior in a VPC is that it it's routed via
device-number-0, but according to entries in the main table it will be
routed via the secondary ENI as the most direct path to the
destination.  This is consistent with Amazon Linux 2's routing
behavior, and for that reason is considered desirable here as well.

Fixes: 0a57c629d ("Route configuration simplification")
Fixes: https://github.com/amazonlinux/amazon-ec2-net-utils/issues/97