removing uncessary IMDS calls and overwrite of configs
We want to remove the IMDS call which is made for the ipv6-cidr block for ipv4 interfaces by checking to see if supports ipv6 first. Then we want to prevent the case where IMDS is throttled, an empty local-ipv4 is returned and the script overwrites an existing config using _install_and_reload. We want to make sure value are not empty before overwriting anything.