Implement per interface IMDS
For requesting IMDS, ec2-net-utils only makes calls through the primray ENI which can lead to throttling by IMDS. We split these calls over the secondary ENIs in order to avoid making too many requests throught the primary interface. If setup_interface cannot use its own interface to call IMDS, use the current primary route, failing to do that use the lowest available secondary. Note, this mechanism is simplistic as there could be edge cases where the only available IMDS interface is not a primary, lowest secondary or itself or if the primary is only able to talk to IMDS.