Secondary IP refresh and IMDS token optimization
This change adds back secondary ip refresh feature which was broken. Now, when a secondary ip is assigned or unassigned it will automatically show up on the host after 60s. Also optimized IMDS token calls such that one token is requested per ENI for a start or refresh cycle. This significantly reduces the number of IMDS token requested per ENI.